Changed post model, controller, posts and users are now connected

+++ server/controllers/feed.js
@@ -4,6 +4,7 @@ const path = require('path')
 const { validationResult } = require('express-validator/check')
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -17,6 +18,7 @@ exports.getPosts = (req, res, next) => {
         .limit(perPage)
     })
     .then(posts => {
+      console.log(posts)
       res.status(200).json({
         message: 'Posts fetched!!',
         posts: posts,
@@ -32,43 +34,51 @@ exports.getPosts = (req, res, next) => {
 }
 
 exports.createPost = (req, res, next) => {
-  const errors = validationResult(req)
-  if(!errors.isEmpty()){
-    const error = new Error('Validation failed')
-    error.statusCode = 422
-    throw error
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    const error = new Error('Validation failed, entered data is incorrect.');
+    error.statusCode = 422;
+    throw error;
   }
-  console.log(req.file)
   if (!req.file) {
-    const error = new Error('No image provided')
-    error.statusCode = 422
-    throw error
+    const error = new Error('No image provided.');
+    error.statusCode = 422;
+    throw error;
   }
   const imageUrl = req.file.path.replace("\\" ,"/");
-  const { title, content } = req.body
+  const title = req.body.title;
+  const content = req.body.content;
+  let creator;
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: {
-      name: 'Niz'
-    }
-  })
-  post.save()
+    creator: req.userId
+  });
+  post
+    .save()
     .then(result => {
-      console.log(result)
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then(result => {
       res.status(201).json({
-        message: 'Post created successfully',
-        post: result
-      })
+        message: 'Post created successfully!',
+        post: post,
+        creator: { _id: creator._id, name: creator.name }
+      });
     })
     .catch(err => {
-      if(!err.statusCode){
-        err.statusCode = 500
+      if (!err.statusCode) {
+        err.statusCode = 500;
       }
-      next(err)
-    })
-}
+      next(err);
+    });
+};
 
 exports.getPost = (req, res, next) => {
   const postId = req.params.postId;
@@ -120,6 +130,11 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 422
         throw error
       }
+      if(post.creator.toString() !== req.userId){
+        const error = new Error('Post not found')
+        error.statusCode = 403
+        throw error
+      }
       if(imageUrl !== post.imageUrl){
         clearImage(post.imageUrl)
       }
@@ -151,8 +166,20 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 422
         throw error
       }
+      if(post.creator.toString() !== req.userId){
+        const error = new Error('Post not found')
+        error.statusCode = 403
+        throw error
+      }
       clearImage(post.imageUrl)
       return Post.findByIdAndRemove(postId)
+    })
+    .then(result => {
+      return User.findById(req.userId)
+    })
+    .then(user => {
+      user.posts.pull(postId)
+      return user.save()
     })
     .then(result => {
       console.log(result)
